requirements-base.txt
ultralytics==8.4.14
transformers==5.1.0
pycocotools>=2.0.7
opencv-python>=4.8
numpy
pandas
tqdm
pyyaml
Pillow
